Fix Memory leak in SSR
The global ERROR object keeps references to the errors objects.
This prevent the garbage collector from freeing the memory.

This fix simply check if a window exists (if yes we are running in the browser) else we are in SSR.
Do not add the error object to the ERROR const global variable to allow GC cleanup.