Feat validator component (#1625)
## Overview

Typically vee-validate is mainly used using the `v-validate` directive which is fine for most use cases. This PR adds a different approach to using the vee-validate library.

It adds a `ValidationProvider` component that handles validation for fields using scoped slots. This makes the slot scope isolated and removes the complexities of using flags and error bag.

```html
<ValidationProvider rules="required">
  <template slot-scope="{ errors, flags, classes }">
    <input type="text" v-model="val" :class="classes">
    <span>{{ errors[0] }}</span>
  </template>
</ValidationProvider>
```

This makes binding classes and flags easier, also the error in the slot scope is a normal array containing the error messages.

It also removes the need to specify a field name. And you no longer have to install vee-validate as a plugin, you may now import the component where the validation is needed and use it there. The only issue that it is more verbose than using the directive.

## Checklist

- [x] validate HTML elements.
- [x] validate Vue components.
- [x] initial value validation.
- [x] handle v-model custom configurations.
- [x] track external changes of the values.
- [x] full flags support.
- [x] composability with other components.
- [x] support target based rules.
- [x] custom events support.
- [x] write tests.

## ✔ Issues closed

closes #1461, closes #1437, closes #1154,