Feat interaction modes (#1911)
🔎 __Overview__

Checklist:
- [x] Implementation.
- [x] Docs.

This PR adds the ability to specify an interaction mode as formalized by [simple-vue-validator](http://simple-vue-validator.magictek.cn/#m_modes).

There are 4 modes implemented for you by default:

- aggressive: this is the default mode and it always validates on `input`.
- passive: doesn't validate on any event, so you can only validate manually.
- lazy: validate on `change`.
- eager: validates on `change` for the first time, then if the field is invalid it will validate on `input`. If the field turns valid again it will go back to validating on `change`. This was a popular request by many developers #1426.

You can customize the mode for each validation provider using the `mode` prop:

```vue
<ValidationProvider rules="required|email" mode="eager">
  <div slot-scope="{ errors }">
    <input v-model="val">
    <span>{{ errors[0] }}</span>
  </div>
</ValidationProvider>
```

The vee-validate plugin has a new method `setMode` which will be used to provide custom modes and changing the mode dynamically as needed.

```js
import VeeValidate from 'vee-validate';

// set all future providers default mode to passive.
// does not affect existing ones.
VeeValidate.setMode('passive');
```

An interaction mode in vee-validate is a function that receives a context similar to the slot-scope props which can be used to customize the behavior of the validation. The function should return an object containing:

- on: Array of event names as strings, those events will be used as triggers for validation.
- debounce: A number to control the validation trigger frequency.

Here is a small example:

```js
import VeeValidate from 'vee-validate';

VeeValidate.setMode('betterEager', ({ errors }) => {
   // Become slightly aggressive if the field is invalid.
  if (errors.length) {
    return {
      on: ['input'],
      debounce: 400
    };
  }

  // validate after 2 seconds of leaving the field 
  return {
    on: ['change'],
    debounce: 2000
  };
});
```

Currently, the mode receives an object with those properties:

- errors: List of error messages.
- flags: the field flag object.
- value: the last validated field value.

These properties may change in the future to make it more powerful. One thing is to make sure that the browser/components emit the configured events properly.

__This feature will deprecate the `events` config but it will remain with a deprecation warning for a couple of releases.__

✔ __Issues affected__

list of issues formatted like this:

closes #1426, closes #1488, closes #501