fix: enable pass args to the i18n messages (#2009)
Flat array params for detect one or more params

🔎 __Overview__

As reported in the issue #1607, vue-i18n default adapter has a limitation. When multiple params are required (e.g.: min, max) both are sended like a only one array. This is a weird behavior, because the pattern adopted (and also explained by some developers that use vue-i18n) is that each value is an argument. Vue-i18n lib has an abstraction for array-like objects thats works fine.