feat: optimize finding fields by id (#2020)
🔎 __Overview__

This PR improves the performance of finding fields by id, using a plain object as a map to link ids to items faster than by looping over the array and matching.

In my profiling of large forms with many fields with validation, the repeated looping over the fields triggered by `findField()` in `src/directive.js` had a serious performance impact that showed up as hundreds of milliseconds in my profiling. 

This simple optimization removes that bottleneck.