Default with field translation? (#2687)
I stumbled above this origin example and used it with vue-i18n plugin. I was wondering why the localization worked almost, but the field didn't get translated. In the other sandbox for LingUI i found what i was searching for and added this line for the fields to the defaultMessage function. It think i would help others.